+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ 1.2.1 ] - 2022-11-12
+## [ 1.3.0 ] - 2023-05-11
+### Added
+- Added switch to enable or disable compare input/output content and set header field. 
+  In many cases the server don't let you compare it (new by default=false).
+
+### Fixed
+- Issues resource handling.
 
 ## [ 1.2.0 ] - 2022-11-12
 ### Added